+++ packages/communication/opc-ethernet-ip-bridge/src/opc-server/index.ts
@@ -89,7 +89,12 @@ export const addTag = async (
 
                 await addTag(server, key, structure?.[key], () => {
                     return getter(key) || defaultValue
-                }, (value) => setter(value, key), undefined, rootObject);
+                }, (value) => {
+                    if(value.BYTES_PER_ELEMENT != undefined) value = Array.from(value);
+
+                    setter(value, key)
+
+                }, undefined, rootObject);
             }
 
             // await server.addVariable(key, getOPCType(structure?.[key] || ''), )
@@ -141,6 +146,9 @@ export const addTag = async (
     }
  
     if(dataType != OPC_TYPE.Structure){
-        await server.addVariable(tagname, dataType, isArray, _getter, setter, parent)
+        await server.addVariable(tagname, dataType, isArray, _getter, (value) => {
+            if(value.BYTES_PER_ELEMENT) value = Array.from(value);
+            setter(value)
+        }, parent)
     }
 }