+++ packages/core/command-data/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.15-alpha.186](https://github.com/TheTechCompany/HiveCommand/compare/v1.4.15-alpha.185...v1.4.15-alpha.186) (2023-02-23)
+
+**Note:** Version bump only for package @hive-command/data
+
+
+
+
+
+## [1.4.15-alpha.185](https://github.com/TheTechCompany/HiveCommand/compare/v1.4.15-alpha.184...v1.4.15-alpha.185) (2023-02-23)
+
+**Note:** Version bump only for package @hive-command/data
+
+
+
+
+
 ## [1.4.15-alpha.184](https://github.com/TheTechCompany/HiveCommand/compare/v1.4.15-alpha.183...v1.4.15-alpha.184) (2023-02-22)
 
 **Note:** Version bump only for package @hive-command/data